accessibility-testing/requirements.txt
# Core Testing Framework
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0

# Accessibility Testing
axe-selenium-python==2.1.6
pa11y==0.1.0

# Browser Drivers
webdriver-manager==4.0.1

# Image Processing (for contrast testing)
Pillow==10.1.0
colorsys2==0.1.0

# Reporting
beautifulsoup4==4.12.2
lxml==4.9.3
jinja2==3.1.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6

# Screen Reader Testing Support
pywinauto==0.6.8; sys_platform == 'win32'
